Refactored some pages.

+++ src/systemComponentCard.tsx
@@ -2,18 +2,51 @@ import {ActionIcon, Button, Card} from "@mantine/core";
 import {IconPencil, IconTrashX} from "@tabler/icons-react";
 import {useNavigate, useParams} from "react-router-dom";
 import {usePouch} from "use-pouchdb";
-
+import log from "loglevel";
+type MasterComponents = {
+    _id: string;
+    _rev: string
+    list: { _id: string, name: string }[];
+}
 export default function SystemComponentCard({data, selectComponent}) {
+    const logger = log.getLogger('SystemComponentCard');
     const params = useParams();
     const db = usePouch();
     const navigate = useNavigate();
     const removeComponent = (id) => {
-        db.get(id).then((doc) => {
-            db.remove(doc);
-        }).catch((err) => {
-            console.log(err);
-        });
+        if (params.solutionId) {
+            db.get(id).then((doc) => {
+                db.remove(doc)
+                    .then(() => {
+                        logger.debug('component removed');
+                    })
+                    .catch((err) => {
+                        logger.error(err);
+                    });
+            }).catch((err) => {
+                logger.error(err);
+            })
+        } else {
+            db.get('components').then((doc) => {
+                logger.debug('grabbing master doc', doc);
+                const masterDoc = (doc as unknown) as  MasterComponents;
+                masterDoc.list = masterDoc.list.filter((component) => {
+                    return component._id !== id;
+                });
+                logger.debug('new master doc', masterDoc);
+                db.put(masterDoc)
+                    .then(() => {
+                        logger.debug('master doc updated');
+                    }).catch((err) => {
+                        logger.error(err);
+                });
+            }).catch((err) => {
+                logger.error(err);
+            });
+        }
+
     }
+    logger.debug(data);
     const selected = params.componentId && (params.componentId == data._id);
     /* <NavLink leftSection={<IconStar color={selected? '#FF0': '#00F'}/>} id={data._id} key={data._id}
                         onClick={selectComponent} label={data.name || data._id}/>
